Fix potential XXE vector via BOM detection
This patch fixes a potential XXE vector that occurs when a multibyte XML
string/file is provided to the security scanner under PHP-FPM when threading
is enabled.

php-src patched libxml to work thread-safe in

- https://github.com/php/php-src/commit/de31324c221c1791b26350ba106cc26bad23ace9

which is included in PHP 5.5 starting in 5.5.22 and PHP 5.6 starting in 5.6.6.
In those versions, we now *always* use the libxml checks.

However, for vulnerable PHP versions, we updated the heuristic algorithm to have
it generate a properly encoded `<!ENTITY` string against which to compare:

- Determine file encoding
  - Check for presence of an actual BOM
    - See https://en.wikipedia.org/wiki/Byte_order_mark for BOM => encoding pairs
  - If no BOM, lookup routine to encode `<?xml` to known encodings,
    and compare that to the current XML string, returning the encoding that
    matches.
  - Otherwise, assume UTF-8
- Find the XML encoding
  - Encode `>` and `encoding="` in the given charset, and see if the latter
    occurs before the former. If so:
    - encode the `"` character, and do get the `substr` from where `encoding="`
      completes to the next `"` character.
    - that string becomes the XML encoding; strip any null bytes and return it.
  - if no `encoding` set, use the file encoding.
- Using the detected encoding, pass encode the string `<!ENTITY`, and use that
  value for the `strpos()` heuristic check.

In point of fact, the patch will use both the XML declared encoding as well as
the file encoding to perform the heuristic check; in many cases, the file
encoding may trump the declared encoding.